+++ src/main/java/team7/inplace/crawling/application/YoutubeCrawlingService.java
@@ -7,7 +7,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import team7.inplace.crawling.application.dto.RawVideoInfo;
 import team7.inplace.crawling.client.YoutubeClient;
-import team7.inplace.crawling.repository.YoutubeChannelRepository;
+import team7.inplace.crawling.persistence.YoutubeChannelRepository;
 
 @Slf4j
 @Service
