+++ src/main/java/team7/inplace/crawling/persistence/YoutubeChannelRepository.java
@@ -1,4 +1,4 @@
-package team7.inplace.crawling.repository;
+package team7.inplace.crawling.persistence;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import team7.inplace.crawling.domain.YoutubeChannel;
